+++ CumulusMX/Stations/DavisCloudStation.cs
@@ -636,7 +636,7 @@ namespace CumulusMX.Stations
 						// reset the accumulated snow depth(s)
 						for (int i = 0; i < Snow24h.Length; i++)
 						{
-							Snow24h[i] = null;
+							Snow24h[i] = LaserDepth[i].HasValue ? 0 : null;
 						}
 
 						snowhourdone = true;