+++ CumulusMX/Stations/TempestStation.cs
@@ -137,7 +137,7 @@ namespace CumulusMX.Stations
 					// reset the accumulated snow depth(s)
 					for (var i = 0; i < Snow24h.Length; i++)
 					{
-						Snow24h[i] = null;
+						Snow24h[i] = LaserDepth[i].HasValue ? 0 : null;
 					}
 
 					snowhourdone = true;