+++ CumulusMX/Stations/WMR200Station.cs
@@ -1433,7 +1433,7 @@ namespace CumulusMX.Stations
 				// reset the accumulated snow depth(s)
 				for (var j = 0; j < Snow24h.Length; j++)
 				{
-					Snow24h[j] = null;
+					Snow24h[j] = LaserDepth[j].HasValue ? 0 : null;
 				}
 
 				snowhourdone = true;